+++ app/Http/Controllers/WatchlistItemController.php
@@ -4,15 +4,15 @@ namespace App\Http\Controllers;
 
 use App\Models\WatchlistItem;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class WatchlistItemController extends Controller
 {
-    public function index()
-    {
-        return view('watchlistItems.index', ['watchlistItems' => WatchlistItem::all()]);
-    }
     public function destroy(WatchlistItem $watchlistItem)
     {
+        if ($watchlistItem->watchlist()->user()->id !== Auth::id()) {
+            return redirect()->back();
+        };
         $watchlistItem->delete();
         return redirect()->back();
     }
